LU-20497 efalnd: Add response cookie to TX to detect stale responses
When a TX is aborted by the daemon and returned to the pool, its slot
can be reused for a new request. If a stale GETR_ACK or completion from
the original request arrives after we completed the Tx, we fetch a Tx
using the index stored in the cookie not used for the message anymore.
This can lead to undefined behaviors since the Tx can be freed but also
can already be used for another message.

Fix by replacing the boolean waiting_resp with an atomic resp_id that
combines a 31-bit resp_counter (incremented per request) and a 1-bit
validity flag. The wire cookie sent to the remote peer embeds the full
resp_id value. On response arrival, it atomically validates the resp_id
matches AND clears validity in a single cmpxchg - eliminating any race
window between validation and claim.

This makes the stale response detection atomic: if the TX was aborted
(xchg in abort clears valid first) or recycled (resp_id changed), the
response handler's cmpxchg fails and the response is dropped.

Test-Parameters: trivial
Signed-off-by: Yonatan Nachum <ynachum@amazon.com>
Change-Id: I78b702995ddfea1c9ce8159b2cbb65255e3a1188
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/67876
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Tom Sela <tomsela@amazon.com>
Reviewed-by: Michael Margolin <mrgolin@amazon.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>