LU-16096 tgt: improve messages for reply_data
Due to the format change in reply_data for WBC, it is not possible
to downgrade the MDT filesystem to an older release after mounting
it with 2.16.0 and later.

On the path to fixing that, improve the error messages printed by
tgt_reply_data_init() so that it is possible to see why the mount
is failing.  The next step is to avoid upgrading the replay_data
file if it is not necessary (no clients mounting with WBC enabled).

Test-Parameters: trivial testlist=conf-sanity env=ONLY=32
Fixes: bbf0017fdea5 ("LU-16096 recovery: upgrade reply data after recovery finish")
Signed-off-by: Andreas Dilger <adilger@dilger.ca>
Change-Id: Idd176031fef99b16de9278f05516dc45e817a819
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/49939
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Vitaliy Kuznetsov <vkuznetsov@ddn.com>
Reviewed-by: Qian Yingjin <qian@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>