LU-13805 osc: Don't include lock for srvlock
When doing server side locking, it doesn't make sense to do
the 'search for covering lock and send it to the server'
step when building an RPC, because we will not use that
lock.

This can disguise issues on the client, because prolonging
a lock is supposed to let a client avoid eviction if it is
still doing IO under the lock, but we are not.  This can
result in delaying an eviction which should be occurring
because the client can't give the lock back.

Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: I6957925bf2d8b7be2340469337906a94a758953d
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/50067
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>