LU-930 doc: move lnet pages, remove old docs
Purge ancient lnet documentation. It doesn't seem to
have been updated in at least 20 years, based on the
git history.

In it's place, put the lnetctl and lst man pages. This
is a more sensible place for the lnet man pages to
live. This patch is preliminary to them being split
up and updated.

Test-Parameters: trivial
Signed-off-by: Timothy Day <timday@amazon.com>
Change-Id: I33a3f38bd0f6803406a235f2177bab21dbf7546b
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/50387
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>