LU-930 misc: update MAINTAINERS addresses
Update email addresses for various maintainers, and remove those
people who are no longer working on Lustre.

Change "M:" to "R:", since "M:" means "Mail to" and not "Maintainer"
as I thought.  Use "R:" for "Reviewer", and remove other tags that
we don't want in this file for Lustre.

Test-Parameters: trivial
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I3e73963b08181154fa48f308cb3d1d0a533ebbe5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/52025
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>