LU-16655 scrub: change sf_uuid to guid_t
Change the type of sf_uuid from uuid_t to guid_t.  The sizes are
identical, but the benefit is that guid_t is usable in userspace
with the <linux/uuid.h> header, unlike uuid_t.

Change the accessors to use the corresponding guid_*() functions,
but no functional changes are needed.

Test-Parameters: trivial
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I3048cfee20a5e4ea2c0c2203d22eb76c1437577b
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/50496
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>