LU-5896 mdt: allow release after blocking migrations
We now pass the new and old dataversion of the file after a blocking
migration to the kernel on the client to the MDS via an extension to
the MDS_SWAP_LAYOUTS RPC. Two data versions are added to the
mdc_swap_layouts part of this RPC.

These new fields are checked on the MDS only if the new flag
SWAP_LAYOUTS_WITH_DV12 is set in mdc_swap_layouts::msl_flags
which was previously set to 0 by the client, so no change in behavior
for old clients. Old servers will ignore the newly-added DV members in
the request as the structure doesn't contain it.

We can then transfer the new dataversion from the client
to the MDS through this extension. The MDD layer can
then update the dataversion accordingly on the server.

Add regression test sanity-hsm 607e.

Test-Parameters: testlist=sanity-hsm env=ONLY=607,ONLY_REPEAT=15
Change-Id: Id5d808d3f3bf64982dce2f930d9c5b5e93dd0251
Signed-off-by: Etienne AUJAMES <eaujames@ddn.com>
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/49377
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Robert Read <rread@thelustrecollective.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Timothy Day <tday141@gmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>